Initial Commit home screen
This is the initial commit of the home screen. The options buttons have
been placed (but not properly sized), the status and list widgets still
need to be populated as well as the button actions defined and
connected.

Also added the demo file from matplotlib to import graphs into Qt
widgets, copyrights in the comments of the script